start loading the BIOS
I've had this in the pocket for a while but never actually committed
it because the ALU and other parts of the CPU needed a lot of work. Now
that the CPU is starting to become more accurate, I feel comfortable
putting this here in preparation for software interrupts.

At the moment, we only load the BIOS into memory. We don't actually do
anything with it yet.